+++ lib/ncs_navigator/warehouse/transformers/enum_transformer.rb
@@ -37,7 +37,7 @@ module NcsNavigator::Warehouse::Transformers
     # @return [void]
     def transform(status)
       enum.each do |record|
-        apply_psu_if_necessary(record)
+        apply_global_values_if_necessary(record)
         if record.valid?
           log.debug("Saving valid record #{record_ident record}.")
           begin
@@ -74,10 +74,16 @@ module NcsNavigator::Warehouse::Transformers
         rec.class.name.demodulize, rec.class.key.first.name, rec.key.try(:first).inspect]
     end
 
-    def apply_psu_if_necessary(record)
-      if record.respond_to?(:psu_id=) && record.respond_to?(:psu_id)
-        unless record.psu_id
-          record.psu_id = @configuration.navigator.psus.first.id
+    def apply_global_values_if_necessary(record)
+      {
+        :psu_id => @configuration.navigator.psus.first.id,
+        :recruit_type => @configuration.navigator.recruitment_type_id
+      }.each do |attr, value|
+        setter = :"#{attr}="
+        if record.respond_to?(setter) && record.respond_to?(attr)
+          unless record.send(attr)
+            record.send(setter, value)
+          end
         end
       end
     end