+++ spec/ncs_navigator/warehouse/transformers/enum_transformer_spec.rb
@@ -6,6 +6,7 @@ module NcsNavigator::Warehouse::Transformers
       include ::DataMapper::Resource
 
       property :psu_id, String
+      property :recruit_type, String
       property :id, Integer, :key => true
       property :name, String, :required => true
     end
@@ -48,6 +49,18 @@ module NcsNavigator::Warehouse::Transformers
 
         subject.transform(transform_status)
         records.collect(&:psu_id).should == %w(20000030 20000042 20000030)
+      end
+
+      it 'automatically sets recruit_type if necessary' do
+        records[2].recruit_type = '1'
+
+        records.each do |m|
+          m.should_receive(:valid?).and_return(true)
+          m.should_receive(:save).and_return(true)
+        end
+
+        subject.transform(transform_status)
+        records.collect(&:recruit_type).should == %w(3 3 1)
       end
 
       describe 'with an invalid instance' do